requirements.txt
# Core dependencies
google-generativeai>=0.3.0
tavily-python>=0.7.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
dataclasses-json>=0.6.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Production deployment
gunicorn>=21.0.0
uvicorn>=0.23.0
fastapi>=0.100.0

# Monitoring and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.40.0

# Caching and performance
redis>=4.6.0
celery>=5.3.0

# Security
cryptography>=41.0.0
passlib>=1.7.4

# API documentation
pydantic-settings>=2.0.0 